# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Simple ID Validation API",
    "description": "A simple API that accepts a plain text string ID and validates it",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/validate": {
      "post": {
        "summary": "Validate ID (plain text)",
        "description": "Accepts a plain text string ID in the request body. Returns error if input is not a valid string.",
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": "abc123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valid string ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Valid input"
                    },
                    "id": {
                      "type": "string",
                      "example": "abc123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - not a string or empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid input: id must be a non-empty string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/validate-json": {
      "post": {
        "summary": "Validate ID (JSON)",
        "description": "Accepts a JSON body with an 'id' field. Returns error if id is not a string.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "abc123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valid string ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Valid input"
                    },
                    "id": {
                      "type": "string",
                      "example": "abc123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - id is not a string or empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid input: id must be a string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}